backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
